refactor(run-types): keep the whole credit-card format in one module
